Move current analytical object specifc functions to separate module
